Gujrati_Agent/requirements.txt
requests==2.32.3
numpy==1.26.4
python-dotenv==1.0.1
openai-whisper==20231117
vosk==0.3.45
gTTS==2.5.3
sentence-transformers==3.2.1
faiss-cpu==1.8.0
sounddevice==0.4.6
soundfile==0.12.1
